pyproject.toml
[project]
name = "sandeco-com-br"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "flask-mail>=0.10.0",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.0",
]
